Integrate introduced changes into humanode-runtime
