+++ crates/humanode-runtime/Cargo.toml
@@ -28,6 +28,7 @@ pallet-evm-accounts-mapping = { path = "../pallet-evm-accounts-mapping", default
 pallet-humanode-session = { path = "../pallet-humanode-session", default-features = false }
 pallet-pot = { path = "../pallet-pot", default-features = false }
 pallet-token-claims = { path = "../pallet-token-claims", default-features = false }
+pallet-token-wrapper = { path = "../pallet-token-wrapper", default-features = false }
 pallet-vesting = { path = "../pallet-vesting", default-features = false }
 precompile-bioauth = { path = "../precompile-bioauth", default-features = false }
 precompile-currency-swap = { path = "../precompile-currency-swap", default-features = false }
@@ -184,6 +185,7 @@ std = [
   "pallet-sudo/std",
   "pallet-timestamp/std",
   "pallet-token-claims/std",
+  "pallet-token-wrapper/std",
   "pallet-transaction-payment-rpc-runtime-api/std",
   "pallet-transaction-payment/std",
   "pallet-utility/std",
@@ -252,6 +254,7 @@ try-runtime = [
   "pallet-sudo/try-runtime",
   "pallet-timestamp/try-runtime",
   "pallet-token-claims/try-runtime",
+  "pallet-token-wrapper/try-runtime",
   "pallet-transaction-payment/try-runtime",
   "pallet-utility/try-runtime",
   "pallet-vesting/try-runtime",
